7-1+1=7/src/main/java/domain/usecase/card/editBlocker/EditBlockerUseCase.java
package domain.usecase.card.editBlocker;

import domain.adapter.card.CardRepository;
import domain.model.card.Card;

public class EditBlockerUseCase {
    private CardRepository cardRepository;

    public EditBlockerUseCase(CardRepository cardRepository) {
        this.cardRepository = cardRepository;
    }

    public void execute(EditBlockerInput input, EditBlockerOutput output) {
        Card card = cardRepository.findById(input.getCardId());
        card.editBlocker(input.getBlocker());
        cardRepository.save(card);

    }
}
